cleaned imports, replaced main.py with previous main_copy.py
